Add notice about version selected for packages from Inqlude

+++ test/install-action-from-inqlude-repository.sh
@@ -1,21 +1,25 @@
 #!/usr/bin/env bash
 
-name="install-inqlude"
+name="packages-from-inqlude"
 fails=0
 i=0
 offline=$1
-tests=`ls ${name}/*${offline}.json | wc -l`
+tests=`ls "${name}"/*"${offline}".json | wc -l`
 echo "1..${tests##* }"
+if [ "${tests##* }" == "0" ]; then
+  exit 255
+fi
+
 for input in ${name}/*${offline}.json
 do
   i=$((i+1))
   if ! ../qompoter.sh install --no-color --file "$input" --repo qompoter-repo | diff -u - "${input%.json}.expected" \
     || ! find vendor -maxdepth 4 | grep -v ".git" | LC_ALL=C sort | diff -u - "${input%.json}.vendor.expected"
   then
-    echo "not ok $i - $input"
+    echo "not ok $i - $(echo $input | tr '-' ' ' | sed 's/.json//')"
     fails=$((fails+1))
   else
-    echo "ok $i - $input"
+    echo "ok $i - $(echo $input | tr '-' ' ' | sed 's/.json//')"
   fi
   rm -rf vendor
 done
